Correction of typing and calculations
Correction of the calculation of Shutter Speed ​​Value; Correction of the altitude value in the GPS; Addition of LightSource and FileSource type.

+++ Types/GPSInfo.cs
@@ -14,7 +14,7 @@ namespace ExifPhotoReader.Types
         public LongitudeRef LongitudeRef { get; set; }
         public double Longitude { get; set; }
         public AltitudeRef AltitudeRef { get; set; }
-        public double Altitude { get; set; }
+        public float Altitude { get; set; }
         public DateTime TimeStamp { get; set; }
         public string Satellites { get; set; }
         public Status Status { get; set; }
